Drafting general design OpenSEMBA/#3
