+++ src/time_integrator/lserk4.py
@@ -3,4 +3,4 @@
 class LSERK4:
 
     def __init__(self, case, spatial_discretization):
-        pass #TODO
+        print ("TBD") #TODO
